pyproject.toml
[tool.poetry]
name = "sharknoise-task-manager"
version = "0.7.1"
description = ""
authors = ["sharknoise <sharknoise@gmail.com>"]

packages = [
  { include = "task_manager" },
]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.12"
psycopg2-binary = "^2.8.5"
django-bootstrap4 = "^2.2.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
python-dotenv = "^0.18.0"
dj-database-url = "^0.5.0"
django-crispy-forms = "^1.12.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.0"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"